corrected third integrals and up requiring division of constant in expression, modified driver to display float precision 2 for conciseness in demo, changed indefinite limit accessors to float params to avoid truncation error

+++ Polynomial.h
@@ -74,8 +74,8 @@ public:
 	friend std::istream& operator>>(std::istream&, Polynomial&);
 
 	// Upper and lower limit acc/mut functions for definite integration
-	void sUpper(int val) { upper = val; };
-	void sLower(int val) { lower = val; };
+	void sUpper(float val) { upper = val; };
+	void sLower(float val) { lower = val; };
 	float gUpper() { return upper; };
 	float gLower() { return lower; };
 };
